+++ services/DuplicatorService.php
@@ -39,6 +39,7 @@ class DuplicatorService
     {
         $this->targetPage->visibility_groups = $this->sourcePage->visibility_groups;
         $this->targetPage->visibility_languages = $this->sourcePage->visibility_languages;
+        $this->targetPage->editors = $this->sourcePage->editors;
 
         foreach ($this->sourcePage->attributes as $attrKey => $attrValue) {
             if ($attrKey !== 'id') {