+++ modules/template/services/TemplateInstanceRendererService.php
@@ -63,8 +63,6 @@ class TemplateInstanceRendererService
      */
     public function render(): string
     {
-        $this->templateInstance->template->registerResources();
-
         if (self::inEditMode() && PagePermissionHelper::canEdit($this->templateInstance->page)) {
             $this->ignoreCache();
         }