+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 - Fix #329: Update module resources path
 - Enh #466: Specific editors for template page
 - Enh #467: Add url property for file template element
+- Enh #468: Improved Module Test GitHub Actions
 
 1.12.1 (September 6, 2025)
 --------------------------